Fixed 'User has no Permissions' error
