+++ src/main/java/org/openjfx/API2Semestre/App.java
@@ -8,6 +8,7 @@ import javafx.stage.Stage;
 
 import java.io.IOException;
 import java.net.URL;
+import java.util.Optional;
 
 import org.openjfx.api2semestre.authentication.Authentication;
 import org.openjfx.api2semestre.view_controllers.BaseController;
@@ -75,36 +76,38 @@ public class App extends Application {
         baseController.getLb_currentUser().setText("Logado como " + Authentication.getCurrentUser().getNome());
     }
 
-    public static void changeView (String newViewFxmlFile) {
+    public static void changeView (Optional<String> newViewFxmlOptional) {
         try {
 
             loadBase();
 
-            Parent module = loadFXML(newViewFxmlFile);
-            baseController.getAp_content().getChildren().add(module);
+            if (newViewFxmlOptional.isPresent()) {
+                String newViewFxml = newViewFxmlOptional.get();
+                Parent module = loadFXML(newViewFxml);
+                baseController.getAp_content().getChildren().add(module);
+                currentViewFxmlFile = newViewFxml;
+
+                for (View view : ViewsManager.getViews()) {
     
-            for (View view : ViewsManager.getViews()) {
-
-                FXMLLoader viewButtonLoader = new FXMLLoader(App.getFXML("templates/viewButtonTemplate"));
-
-                baseController.getVb_views().getChildren().add(viewButtonLoader.load());
-
-                ViewButtonController viewButtonTemplateController = viewButtonLoader.getController();
-
-                String buttonViewFxmlFile = view.getFxmlFileName();
-
-                viewButtonTemplateController.setView(buttonViewFxmlFile);
-                viewButtonTemplateController.setText(view.getDisplayName());
-
-                if (buttonViewFxmlFile.equals(currentViewFxmlFile)) viewButtonTemplateController.setDisable(false);
-                if (buttonViewFxmlFile.equals(newViewFxmlFile)) viewButtonTemplateController.setDisable(true);
+                    FXMLLoader viewButtonLoader = new FXMLLoader(App.getFXML("templates/viewButtonTemplate"));
+    
+                    baseController.getVb_views().getChildren().add(viewButtonLoader.load());
+    
+                    ViewButtonController viewButtonTemplateController = viewButtonLoader.getController();
+    
+                    String buttonViewFxmlFile = view.getFxmlFileName();
+    
+                    viewButtonTemplateController.setView(buttonViewFxmlFile);
+                    viewButtonTemplateController.setText(view.getDisplayName());
+    
+                    if (buttonViewFxmlFile.equals(currentViewFxmlFile)) viewButtonTemplateController.setDisable(false);
+                    if (buttonViewFxmlFile.equals(newViewFxml)) viewButtonTemplateController.setDisable(true);
+                }
 
             }
 
-            currentViewFxmlFile = newViewFxmlFile;
-
         } catch (Exception ex) {
-            System.out.println("App.changeView() -- Erro ao trocar para a tela '" + newViewFxmlFile + "'");
+            System.out.println("App.changeView() -- Erro ao trocar para a tela '" + newViewFxmlOptional + "'");
             ex.printStackTrace();
         }
     }
